feat(presenter-ksp)!: split Koin generation into presenter-koin-ksp
BREAKING CHANGE: lbcpresenter-ksp no longer generates the Koin module
or honors annotateFactory on its own. Koin projects must apply
ksp(lbcpresenter-koin-ksp) instead (pulls lbcpresenter-ksp
transitively); KSP option names are unchanged.

- presenter-ksp: factory generation only, no Koin output. Exposes
  ReducerSignatureParser, GeneratedFactoryDecorator hook and an open
  ReducerFactoryProcessor so DI processors can extend it
- presenter-koin-ksp: generates generatedReducerFactoryModule and
  annotated factories (@Factory + qualifiers) via KoinFactoryDecorator
- when annotateFactory=true, core processor skips factory generation
  so the Koin processor can emit the annotated variant without
  duplicate files
