+++ compose/presenter/presenter-koin-ksp/.gitignore
@@ -1,0 +1,1 @@
+/build
