+++ compose/presenter/presenter-koin-ksp/build.gradle.kts
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2026 Lunabee Studio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+plugins {
+    kotlin("jvm")
+    id("lunabee.java-library-conventions")
+    id("lunabee.library-publish-conventions")
+}
+
+description = "Koin KSP integration for LBPresenter"
+version = AndroidConfig.LBCPRESENTER_KOIN_KSP_VERSION
+
+dependencies {
+    implementation(libs.kotlinPoet)
+    implementation(libs.kspApi)
+
+    api(projects.compose.presenterKsp)
+
+    testImplementation(libs.junit4)
+    testImplementation(libs.kotlinTest)
+}
