+++ compose/presenter/presenter-koin-ksp/src/main/kotlin/studio/lunabee/compose/presenter/ksp/koin/KoinReducerFactoryProcessor.kt
@@ -1,0 +1,157 @@
+/*
+ * Copyright (c) 2026 Lunabee Studio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package studio.lunabee.compose.presenter.ksp.koin
+
+import com.google.devtools.ksp.processing.CodeGenerator
+import com.google.devtools.ksp.processing.Dependencies
+import com.google.devtools.ksp.processing.KSPLogger
+import com.google.devtools.ksp.processing.PlatformInfo
+import com.google.devtools.ksp.processing.Resolver
+import com.google.devtools.ksp.processing.SymbolProcessor
+import com.google.devtools.ksp.processing.SymbolProcessorEnvironment
+import com.google.devtools.ksp.processing.SymbolProcessorProvider
+import com.google.devtools.ksp.symbol.KSAnnotated
+import com.google.devtools.ksp.validate
+import studio.lunabee.compose.presenter.ksp.ReducerFactoryProcessor
+import studio.lunabee.compose.presenter.ksp.ValidReducerSignature
+
+private const val GenerateKoinModuleOption = "studio.lunabee.presenter.generateKoinModule"
+private const val KoinModulePackageOption = "studio.lunabee.presenter.koinModulePackage"
+private const val AnnotateFactoryOption = "studio.lunabee.presenter.annotateFactory"
+
+class KoinReducerFactoryProcessorProvider : SymbolProcessorProvider {
+    /**
+     * Creates the processor used to generate the Koin bindings of generated reducer factories.
+     */
+    override fun create(environment: SymbolProcessorEnvironment): SymbolProcessor {
+        val annotateFactory = environment.options[AnnotateFactoryOption]?.toBooleanStrictOrNull() == true
+        val koinModuleGenerationRequested = environment.options[GenerateKoinModuleOption]?.toBooleanStrictOrNull() == true
+        val generateKoinModule = koinModuleGenerationRequested && shouldGenerateKoinModuleForCompilation(environment.platforms)
+        if (koinModuleGenerationRequested && !generateKoinModule) {
+            environment.logger.info(
+                "Skipping reducer Koin module generation for this compilation because it cannot aggregate all factories. " +
+                    "A platform compilation will generate the shared module instead.",
+            )
+        }
+        return KoinReducerFactoryProcessor(
+            codeGenerator = environment.codeGenerator,
+            logger = environment.logger,
+            annotateFactory = annotateFactory,
+            generateKoinModule = generateKoinModule,
+            configuredKoinModulePackageName = environment.options[KoinModulePackageOption]?.trim()?.takeIf { it.isNotEmpty() },
+        )
+    }
+}
+
+internal fun shouldGenerateKoinModuleForCompilation(platforms: List<PlatformInfo>): Boolean = platforms.size == 1
+
+/**
+ * Generates the Koin bindings of the reducer factories generated by the lbcpresenter-ksp processor.
+ *
+ * When [annotateFactory] is true, this processor also takes over factory generation (the lbcpresenter-ksp processor
+ * skips it) so the generated factory classes carry [org.koin.core.annotation.Factory] and the Koin qualifiers declared
+ * on injected reducer constructor parameters.
+ */
+internal class KoinReducerFactoryProcessor(
+    codeGenerator: CodeGenerator,
+    logger: KSPLogger,
+    annotateFactory: Boolean,
+    private val generateKoinModule: Boolean,
+    private val configuredKoinModulePackageName: String?,
+) : ReducerFactoryProcessor(
+    codeGenerator = codeGenerator,
+    logger = logger,
+    generateFactories = annotateFactory,
+    factoryDecorator = KoinFactoryDecorator.takeIf { annotateFactory },
+) {
+    private val moduleFileGenerator: KoinModuleFileGenerator = KoinModuleFileGenerator()
+    private val koinModuleSignatures: LinkedHashMap<String, ValidReducerSignature> = linkedMapOf()
+    private val moduleSourcePackageNames: LinkedHashSet<String> = linkedSetOf()
+    private var moduleRootPackageName: String? = null
+    private var hasWarnedAboutMissingKoinModulePackageOption: Boolean = false
+
+    override fun onProcessStart(resolver: Resolver) {
+        if (!generateKoinModule) return
+
+        moduleSourcePackageNames += resolver.getAllFiles().map { it.packageName.asString() }.toList()
+        moduleRootPackageName = resolveModuleRootPackageName(
+            configuredPackageName = configuredKoinModulePackageName,
+            sourcePackageNames = moduleSourcePackageNames.toList(),
+        )
+        warnAboutMissingKoinModulePackageOptionIfNeeded()
+    }
+
+    override fun onProcessEnd(
+        resolver: Resolver,
+        deferred: MutableList<KSAnnotated>,
+    ) {
+        if (!generateKoinModule) return
+
+        koinModuleSignatures.clear()
+        // KSP incremental runs may not surface every annotated reducer in getSymbolsWithAnnotation().
+        // Rebuilding the aggregating module from all visible source files keeps the shared Koin module stable.
+        resolver.getAllFiles()
+            .asSequence()
+            .flatMap { file -> parser.annotatedReducerDeclarations(file) }
+            .forEach { declaration ->
+                if (!declaration.validate()) {
+                    deferred += declaration
+                    return@forEach
+                }
+
+                val signature = buildValidSignature(declaration) ?: return@forEach
+                koinModuleSignatures[signature.factoryClassName.canonicalName] = signature
+            }
+    }
+
+    override fun finish() {
+        if (!shouldGenerateKoinModuleAtFinish(generateKoinModule, koinModuleSignatures.values)) return
+
+        val fileSpec = moduleFileGenerator.generateKoinModule(
+            signatures = koinModuleSignatures.values.toList(),
+            moduleRootPackageName = moduleRootPackageName
+                ?: commonPackageName(koinModuleSignatures.values.map { it.packageName }),
+        )
+        // The shared Koin module needs full-source invalidation so KSP does not drop it on unrelated incremental reruns.
+        writeGeneratedFile(fileSpec, Dependencies.ALL_FILES)
+    }
+
+    private fun warnAboutMissingKoinModulePackageOptionIfNeeded() {
+        if (configuredKoinModulePackageName != null) return
+        if (hasWarnedAboutMissingKoinModulePackageOption) return
+        logger.warn(
+            "KSP option '$KoinModulePackageOption' is not set. Falling back to package inference from source files for " +
+                "generatedReducerFactoryModule. Add the option to make the generated Koin module package explicit and stable.",
+        )
+        hasWarnedAboutMissingKoinModulePackageOption = true
+    }
+}
+
+internal fun shouldGenerateKoinModuleAtFinish(
+    generateKoinModule: Boolean,
+    signatures: Collection<ValidReducerSignature>,
+): Boolean = generateKoinModule && signatures.isNotEmpty()
+
+internal fun resolveModuleRootPackageName(
+    configuredPackageName: String?,
+    sourcePackageNames: List<String>,
+): String {
+    configuredPackageName?.takeIf { it.isNotBlank() }?.let { return it }
+
+    val normalizedSourcePackageNames = sourcePackageNames.filter { it.isNotBlank() }
+    return normalizedSourcePackageNames.firstOrNull()?.let { commonPackageName(normalizedSourcePackageNames) } ?: ""
+}
