+++ compose/presenter/presenter-koin-ksp/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,0 +1,1 @@
+studio.lunabee.compose.presenter.ksp.koin.KoinReducerFactoryProcessorProvider
