+++ compose/presenter/presenter-ksp/src/test/kotlin/studio/lunabee/compose/presenter/ksp/ReducerFactoryFileGeneratorTest.kt
@@ -194,220 +194,6 @@ class ReducerFactoryFileGeneratorTest {
     }
 
     @Test
-    fun generate_koin_module_preserves_qualifiers_test() {
-        val actionType = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerAction")
-        val qualifierType = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerQualifier")
-        val validSignature = validator.validate(
-            RawReducerSignature(
-                packageName = "studio.lunabee.compose.demo.presenter.timer",
-                reducerClassName = ClassName("studio.lunabee.compose.demo.presenter.timer", "QualifiedTimerReducer"),
-                uiStateTypeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerUiState"),
-                navScopeTypeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerNavScope"),
-                actionTypeName = actionType,
-                constructorVisibility = Visibility.Public,
-                constructorParameters = listOf(
-                    RawReducerParameter(
-                        name = "coroutineScope",
-                        typeName = ClassName("kotlinx.coroutines", "CoroutineScope"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                    RawReducerParameter(
-                        name = "emitUserAction",
-                        typeName = LambdaTypeName.get(parameters = arrayOf(actionType), returnType = Unit::class.asTypeName()),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                    RawReducerParameter(
-                        name = "apiClient",
-                        typeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerApiClient"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                        qualifier = KoinQualifier.Named("api"),
-                    ),
-                    RawReducerParameter(
-                        name = "qualifierScopedDependency",
-                        typeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerQualifierScopedDependency"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                        qualifier = KoinQualifier.Typed(qualifierType),
-                    ),
-                ),
-            ),
-        )
-
-        val generatedSource = generator.render(
-            generator.generateKoinModule(
-                signatures = listOf(validSignature),
-                moduleRootPackageName = "studio.lunabee.compose.demo",
-            ),
-        )
-
-        assertTrue(generatedSource.contains("get(qualifier = named(\"api\"))"))
-        assertTrue(generatedSource.contains("qualifier ="))
-        assertTrue(generatedSource.contains("named<TimerQualifier>()"))
-    }
-
-    @Test
-    fun generate_koin_module_file_test() {
-        val timerActionType = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerAction")
-        val timerSignature = validator.validate(
-            RawReducerSignature(
-                packageName = "studio.lunabee.compose.demo.presenter.timer",
-                reducerClassName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerReducer"),
-                uiStateTypeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerUiState"),
-                navScopeTypeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerNavScope"),
-                actionTypeName = timerActionType,
-                constructorVisibility = Visibility.Public,
-                constructorParameters = listOf(
-                    RawReducerParameter(
-                        name = "coroutineScope",
-                        typeName = ClassName("kotlinx.coroutines", "CoroutineScope"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                    RawReducerParameter(
-                        name = "emitUserAction",
-                        typeName = LambdaTypeName.get(parameters = arrayOf(timerActionType), returnType = Unit::class.asTypeName()),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                    RawReducerParameter(
-                        name = "injectedParam",
-                        typeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerInjectedParam"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                    RawReducerParameter(
-                        name = "runtimeParam",
-                        typeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerRuntimeParam"),
-                        hasRuntimeAnnotation = true,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                ),
-            ),
-        )
-        val simpleActionType = ClassName("studio.lunabee.compose.demo.presenter.simple", "SimpleAction")
-        val simpleSignature = validator.validate(
-            RawReducerSignature(
-                packageName = "studio.lunabee.compose.demo.presenter.simple",
-                reducerClassName = ClassName("studio.lunabee.compose.demo.presenter.simple", "SimpleReducer"),
-                uiStateTypeName = ClassName("studio.lunabee.compose.demo.presenter.simple", "SimpleUiState"),
-                navScopeTypeName = ClassName("studio.lunabee.compose.demo.presenter.simple", "SimpleNavScope"),
-                actionTypeName = simpleActionType,
-                constructorVisibility = Visibility.Public,
-                constructorParameters = listOf(
-                    RawReducerParameter(
-                        name = "coroutineScope",
-                        typeName = ClassName("kotlinx.coroutines", "CoroutineScope"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                    RawReducerParameter(
-                        name = "emitUserAction",
-                        typeName = LambdaTypeName.get(parameters = arrayOf(simpleActionType), returnType = Unit::class.asTypeName()),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                ),
-            ),
-        )
-
-        val fileSpec = generator.generateKoinModule(
-            signatures = listOf(timerSignature, simpleSignature),
-            moduleRootPackageName = "studio.lunabee.compose.demo",
-        )
-        val generatedSource = generator.render(fileSpec)
-
-        assertEquals("studio.lunabee.compose.demo", fileSpec.packageName)
-        assertTrue(generatedSource.contains("public val generatedReducerFactoryModule: Module = module {"))
-        assertTrue(generatedSource.contains("factory { SimpleReducerFactory() }"))
-        assertTrue(generatedSource.contains("factory { TimerReducerFactory(get()) }"))
-    }
-
-    @Test
-    fun generate_annotated_factory_file_test() {
-        val annotatingGenerator = ReducerFactoryFileGenerator(annotateFactory = true)
-        val actionType = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerAction")
-        val qualifierType = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerQualifier")
-        val validSignature = validator.validate(
-            RawReducerSignature(
-                packageName = "studio.lunabee.compose.demo.presenter.timer",
-                reducerClassName = ClassName("studio.lunabee.compose.demo.presenter.timer", "AnnotatedTimerReducer"),
-                uiStateTypeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerUiState"),
-                navScopeTypeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerNavScope"),
-                actionTypeName = actionType,
-                constructorVisibility = Visibility.Public,
-                constructorParameters = listOf(
-                    RawReducerParameter(
-                        name = "coroutineScope",
-                        typeName = ClassName("kotlinx.coroutines", "CoroutineScope"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                    RawReducerParameter(
-                        name = "emitUserAction",
-                        typeName = LambdaTypeName.get(parameters = arrayOf(actionType), returnType = Unit::class.asTypeName()),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                    RawReducerParameter(
-                        name = "plainDependency",
-                        typeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerPlainDependency"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                    RawReducerParameter(
-                        name = "apiClient",
-                        typeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerApiClient"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                        qualifier = KoinQualifier.Named("api"),
-                    ),
-                    RawReducerParameter(
-                        name = "qualifierScopedDependency",
-                        typeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerQualifierScopedDependency"),
-                        hasRuntimeAnnotation = false,
-                        hasDefault = false,
-                        isVararg = false,
-                        qualifier = KoinQualifier.Typed(qualifierType),
-                    ),
-                    RawReducerParameter(
-                        name = "external",
-                        typeName = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerExternalRuntimeArg"),
-                        hasRuntimeAnnotation = true,
-                        hasDefault = false,
-                        isVararg = false,
-                    ),
-                ),
-            ),
-        )
-
-        val generatedSource = annotatingGenerator.render(annotatingGenerator.generate(validSignature))
-
-        assertTrue(generatedSource.contains("org.koin.core.`annotation`.Factory"))
-        assertTrue(generatedSource.contains("@Factory\npublic class AnnotatedTimerReducerFactory"))
-        assertTrue(generatedSource.contains("@Named(\"api\")"))
-        assertTrue(generatedSource.contains("@TimerQualifier"))
-        // @FactoryArg params stay out of the injectable constructor (they live in the FactoryArgs data class).
-        assertFalse(generatedSource.contains("private val `external`"))
-    }
-
-    @Test
     fun generate_factory_without_annotation_by_default_test() {
         val actionType = ClassName("studio.lunabee.compose.demo.presenter.timer", "TimerAction")
         val validSignature = validator.validate(
@@ -441,19 +227,5 @@ class ReducerFactoryFileGeneratorTest {
 
         assertFalse(generatedSource.contains("org.koin.core.annotation.Factory"))
         assertFalse(generatedSource.contains("@Factory"))
-    }
-
-    @Test
-    fun common_package_name_test() {
-        assertEquals(
-            "studio.lunabee.compose.demo",
-            commonPackageName(
-                listOf(
-                    "studio.lunabee.compose.demo.presenter.timer",
-                    "studio.lunabee.compose.demo.presenter.simple",
-                    "studio.lunabee.compose.demo.navigation",
-                ),
-            ),
-        )
     }
 }
